Add voice control
The voice control is a bit iffy/temperemental still. The wake word is 'R2', so:

`R2... open panels`
`R2... scream`

The exports for the picovoice config are in the `voice-control` directory in
case they purge my account again.

The nano board TX plugs into pin 34 of the Penumbra Shadow MD board.